feat48/move the tournament related code to its own app (#49)
* feat48/move the tournament related code to its own app
* Todo: replace the game endpoint with pong and remove it from beePong when the server side pong is merged to main
