+++ srcs/backend/beePong/templates/beePong/home.html
@@ -7,7 +7,7 @@
 					<p>The mighty Pong contest!*</p>
 					<p>Play solo against our vicious AI or gather friends and let the tournament decide who's the queen bee 🐝</p>
 				</div>
-				<button class="button--primary button--with-icon button--full-width" onclick="navigate(`{% url 'beePong:tournament' %}`, '/tournament')">
+				<button class="button--primary button--with-icon button--full-width" onclick="navigate(`{% url 'tournament:tournament' %}`, '/tournament')">
 					<span class="button__text">TOURNAMENT</span>
 					<img src="/img/arrow.svg" alt="arrow" class="arrow-icon">
 				</button>
