+++ srcs/backend/bpong_project/urls.py
@@ -22,6 +22,7 @@ from django.conf.urls import handler404
 urlpatterns = [
     path("admin/", admin.site.urls),
     path('accounts/', include('accounts.urls')),
+    path('tournament/', include('tournament.urls')),
     path('', include('beePong.urls')),
 ]
