+++ srcs/backend/tournament/serializers.py
@@ -1,22 +1,22 @@
-from rest_framework import serializers
-from .models import Player, Game, Tournament, PlayerTournament
+# from rest_framework import serializers
+# from .models import Player, Game, Tournament, PlayerTournament
 
-class PlayerSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Player
-        fields = '__all__'
+# class PlayerSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Player
+#         fields = '__all__'
 
-class GameSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Game
-        fields = '__all__'
+# class GameSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Game
+#         fields = '__all__'
 
-class TournamentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tournament
-        fields = '__all__'
+# class TournamentSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = Tournament
+#         fields = '__all__'
 
-class PlayerTournamentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = PlayerTournament
-        fields = '__all__'
+# class PlayerTournamentSerializer(serializers.ModelSerializer):
+#     class Meta:
+#         model = PlayerTournament
+#         fields = '__all__'
