+++ srcs/backend/tournament/templates/tournament/create_tournament.html
