+++ srcs/backend/tournament/templates/tournament/tournament.html
@@ -1,7 +1,7 @@
 <section>
 	<div class="tournament__container">
 		<h1>TOURNAMENTS</h1>
-		<button class="button--primary" onclick="navigate(`{% url 'beePong:create_tournament' %}`)">+ NEW</button>
+		<button class="button--primary" onclick="navigate(`{% url 'tournament:create_tournament' %}`)">+ NEW</button>
 	</div>
 
 	<div class="tournament__grid">
